feat(commands): Add volume and search functionality, update play method
This commit introduces two new features to the discord bot: a 'volume' command for controlling the volume of the music player, and a 'search' command that lets the users search for songs and play them directly. The change also involved updating the 'play' command to streamline the playing process, as it now directly plays the first track from the search results instead of waiting for user input.

In the 'extractor/index.ts' file, the bot now returns an array of tracks from the search results instead of a single track. This change allows the bot to queue up all search result songs, improving user experience when listening to music. The comment in 'netease.ts' was also corrected for grammatical accuracy. Lastly, the version of '@flowease/extractor' has been increased in the 'package-lock.json' file.